chore: split miden_processor::fast into multiple files
