+++ processor/src/fast/basic_block/crypto_ops.rs
