+++ processor/src/fast/mod.rs
@@ -4,26 +4,19 @@ use core::cmp::min;
 use memory::Memory;
 use miden_air::RowIndex;
 use miden_core::{
-    Decorator, DecoratorIterator, EMPTY_WORD, Felt, ONE, Operation, Program, StackOutputs,
-    WORD_SIZE, Word, ZERO,
-    mast::{
-        BasicBlockNode, CallNode, ExternalNode, JoinNode, LoopNode, MastForest, MastNode,
-        MastNodeId, OpBatch, SplitNode,
-    },
+    Decorator, EMPTY_WORD, Felt, Program, StackOutputs, WORD_SIZE, Word, ZERO,
+    mast::{MastForest, MastNode, MastNodeId},
     stack::MIN_STACK_DEPTH,
     utils::range,
 };
 
 use crate::{
     AdviceInputs, AdviceProvider, AsyncHost, ContextId, ErrorContext, ExecutionError, FMP_MIN,
-    ProcessState, SYSCALL_FMP_MIN,
+    ProcessState,
     chiplets::Ace,
     continuation_stack::{Continuation, ContinuationStack},
-    err_ctx,
     fast::{
-        execution_tracer::ExecutionTracer,
-        trace_state::{NodeExecutionState, TraceFragmentContext},
-        tracer::NoopTracer,
+        execution_tracer::ExecutionTracer, trace_state::TraceFragmentContext, tracer::NoopTracer,
     },
 };
 
@@ -33,16 +26,12 @@ pub mod trace_state;
 mod tracer;
 pub use tracer::Tracer;
 
-// Ops
-mod circuit_eval;
-mod crypto_ops;
-mod field_ops;
-mod fri_ops;
-mod horner_ops;
-mod io_ops;
-mod stack_ops;
-mod sys_ops;
-mod u32_ops;
+mod basic_block;
+mod call_and_dyn;
+mod external;
+mod join;
+mod r#loop;
+mod split;
 
 #[cfg(test)]
 mod tests;
@@ -64,12 +53,6 @@ const STACK_BUFFER_SIZE: usize = 6850;
 /// 0's that were generated automatically to keep the stack depth at 16. In practice, if this
 /// occurs, it is most likely a bug.
 const INITIAL_STACK_TOP_IDX: usize = 250;
-
-/// WORD_SIZE, but as a `Felt`.
-const WORD_SIZE_FELT: Felt = Felt::new(4);
-
-/// The size of a double-word.
-const DOUBLE_WORD_SIZE: Felt = Felt::new(8);
 
 /// The number of rows per core trace fragment.
 pub const NUM_ROWS_PER_CORE_FRAGMENT: usize = 1024;
@@ -220,6 +203,12 @@ impl FastProcessor {
 
     // ACCESSORS
     // -------------------------------------------------------------------------------------------
+
+    /// Returns the size of the stack.
+    #[inline(always)]
+    fn stack_size(&self) -> usize {
+        self.stack_top_idx - self.stack_bot_idx
+    }
 
     /// Returns the stack, such that the top of the stack is at the last index of the returned
     /// slice.
@@ -493,690 +482,8 @@ impl FastProcessor {
         })
     }
 
-    // NODE EXECUTORS
+    // DECORATOR EXECUTORS
     // --------------------------------------------------------------------------------------------
-
-    /// Executes a Join node from the start.
-    #[inline(always)]
-    fn start_join_node(
-        &mut self,
-        join_node: &JoinNode,
-        node_id: MastNodeId,
-        current_forest: &Arc<MastForest>,
-        continuation_stack: &mut ContinuationStack,
-        host: &mut impl AsyncHost,
-        tracer: &mut impl Tracer,
-    ) -> Result<(), ExecutionError> {
-        tracer.start_clock_cycle(
-            self,
-            NodeExecutionState::Start(node_id),
-            continuation_stack,
-            current_forest,
-        );
-
-        // Execute decorators that should be executed before entering the node
-        self.execute_before_enter_decorators(node_id, current_forest, host)?;
-
-        continuation_stack.push_finish_join(node_id);
-        continuation_stack.push_start_node(join_node.second());
-        continuation_stack.push_start_node(join_node.first());
-
-        // Corresponds to the row inserted for the JOIN operation added
-        // to the trace.
-        self.increment_clk(tracer);
-
-        Ok(())
-    }
-
-    /// Executes the finish phase of a Join node.
-    #[inline(always)]
-    fn finish_join_node(
-        &mut self,
-        node_id: MastNodeId,
-        current_forest: &Arc<MastForest>,
-        continuation_stack: &mut ContinuationStack,
-        host: &mut impl AsyncHost,
-        tracer: &mut impl Tracer,
-    ) -> Result<(), ExecutionError> {
-        tracer.start_clock_cycle(
-            self,
-            NodeExecutionState::End(node_id),
-            continuation_stack,
-            current_forest,
-        );
-
-        // Corresponds to the row inserted for the END operation added
-        // to the trace.
-        self.increment_clk(tracer);
-
-        self.execute_after_exit_decorators(node_id, current_forest, host)
-    }
-
-    /// Executes a Split node from the start.
-    #[inline(always)]
-    fn start_split_node(
-        &mut self,
-        split_node: &SplitNode,
-        node_id: MastNodeId,
-        current_forest: &Arc<MastForest>,
-        continuation_stack: &mut ContinuationStack,
-        host: &mut impl AsyncHost,
-        tracer: &mut impl Tracer,
-    ) -> Result<(), ExecutionError> {
-        tracer.start_clock_cycle(
-            self,
-            NodeExecutionState::Start(node_id),
-            continuation_stack,
-            current_forest,
-        );
-
-        // Execute decorators that should be executed before entering the node
-        self.execute_before_enter_decorators(node_id, current_forest, host)?;
-
-        let condition = self.stack_get(0);
-
-        // drop the condition from the stack
-        self.decrement_stack_size(tracer);
-
-        // execute the appropriate branch
-        continuation_stack.push_finish_split(node_id);
-        if condition == ONE {
-            continuation_stack.push_start_node(split_node.on_true());
-        } else if condition == ZERO {
-            continuation_stack.push_start_node(split_node.on_false());
-        } else {
-            let err_ctx = err_ctx!(current_forest, split_node, host);
-            return Err(ExecutionError::not_binary_value_if(condition, &err_ctx));
-        };
-
-        // Corresponds to the row inserted for the SPLIT operation added
-        // to the trace.
-        self.increment_clk(tracer);
-
-        Ok(())
-    }
-
-    /// Executes the finish phase of a Split node.
-    #[inline(always)]
-    fn finish_split_node(
-        &mut self,
-        node_id: MastNodeId,
-        current_forest: &Arc<MastForest>,
-        continuation_stack: &mut ContinuationStack,
-        host: &mut impl AsyncHost,
-        tracer: &mut impl Tracer,
-    ) -> Result<(), ExecutionError> {
-        tracer.start_clock_cycle(
-            self,
-            NodeExecutionState::End(node_id),
-            continuation_stack,
-            current_forest,
-        );
-
-        // Corresponds to the row inserted for the END operation added
-        // to the trace.
-        self.increment_clk(tracer);
-
-        self.execute_after_exit_decorators(node_id, current_forest, host)
-    }
-
-    /// Executes a Loop node from the start.
-    #[inline(always)]
-    fn start_loop_node(
-        &mut self,
-        loop_node: &LoopNode,
-        current_node_id: MastNodeId,
-        current_forest: &Arc<MastForest>,
-        continuation_stack: &mut ContinuationStack,
-        host: &mut impl AsyncHost,
-        tracer: &mut impl Tracer,
-    ) -> Result<(), ExecutionError> {
-        tracer.start_clock_cycle(
-            self,
-            NodeExecutionState::Start(current_node_id),
-            continuation_stack,
-            current_forest,
-        );
-
-        // Execute decorators that should be executed before entering the node
-        self.execute_before_enter_decorators(current_node_id, current_forest, host)?;
-
-        let condition = self.stack_get(0);
-
-        // drop the condition from the stack
-        self.decrement_stack_size(tracer);
-
-        // execute the loop body as long as the condition is true
-        if condition == ONE {
-            // Push the loop to check condition again after body
-            // executes
-            continuation_stack.push_finish_loop(current_node_id);
-            continuation_stack.push_start_node(loop_node.body());
-
-            // Corresponds to the row inserted for the LOOP operation added
-            // to the trace.
-            self.increment_clk(tracer);
-        } else if condition == ZERO {
-            // Start and exit the loop immediately - corresponding to adding a LOOP and END row
-            // immediately since there is no body to execute.
-
-            // Increment the clock, corresponding to the LOOP operation
-            self.increment_clk(tracer);
-
-            tracer.start_clock_cycle(
-                self,
-                NodeExecutionState::End(current_node_id),
-                continuation_stack,
-                current_forest,
-            );
-
-            // Increment the clock, corresponding to the END operation added to the trace.
-            self.increment_clk(tracer);
-        } else {
-            let err_ctx = err_ctx!(current_forest, loop_node, host);
-            return Err(ExecutionError::not_binary_value_loop(condition, &err_ctx));
-        }
-        Ok(())
-    }
-
-    /// Executes the finish phase of a Loop node.
-    #[inline(always)]
-    fn finish_loop_node(
-        &mut self,
-        current_node_id: MastNodeId,
-        current_forest: &Arc<MastForest>,
-        continuation_stack: &mut ContinuationStack,
-        host: &mut impl AsyncHost,
-        tracer: &mut impl Tracer,
-    ) -> Result<(), ExecutionError> {
-        // This happens after loop body execution
-        // Check condition again to see if we should continue looping
-        let condition = self.stack_get(0);
-        let loop_node = current_forest[current_node_id].unwrap_loop();
-
-        if condition == ONE {
-            // Add REPEAT row and continue looping
-            tracer.start_clock_cycle(
-                self,
-                NodeExecutionState::LoopRepeat(current_node_id),
-                continuation_stack,
-                current_forest,
-            );
-
-            // Drop the condition from the stack (on the REPEAT instruction)
-            self.decrement_stack_size(tracer);
-
-            continuation_stack.push_finish_loop(current_node_id);
-            continuation_stack.push_start_node(loop_node.body());
-
-            // Corresponds to the REPEAT operation added to the trace.
-            self.increment_clk(tracer);
-        } else if condition == ZERO {
-            // Exit the loop - add END row
-            tracer.start_clock_cycle(
-                self,
-                NodeExecutionState::End(current_node_id),
-                continuation_stack,
-                current_forest,
-            );
-            self.decrement_stack_size(tracer);
-
-            // Corresponds to the END operation added to the trace.
-            self.increment_clk(tracer);
-            self.execute_after_exit_decorators(current_node_id, current_forest, host)?;
-        } else {
-            let err_ctx = err_ctx!(current_forest, loop_node, host);
-            return Err(ExecutionError::not_binary_value_loop(condition, &err_ctx));
-        }
-        Ok(())
-    }
-
-    /// Executes a Call node from the start.
-    #[allow(clippy::too_many_arguments)]
-    #[inline(always)]
-    fn start_call_node(
-        &mut self,
-        call_node: &CallNode,
-        current_node_id: MastNodeId,
-        program: &Program,
-        current_forest: &Arc<MastForest>,
-        continuation_stack: &mut ContinuationStack,
-        host: &mut impl AsyncHost,
-        tracer: &mut impl Tracer,
-    ) -> Result<(), ExecutionError> {
-        tracer.start_clock_cycle(
-            self,
-            NodeExecutionState::Start(current_node_id),
-            continuation_stack,
-            current_forest,
-        );
-
-        // Execute decorators that should be executed before entering the node
-        self.execute_before_enter_decorators(current_node_id, current_forest, host)?;
-
-        let err_ctx = err_ctx!(current_forest, call_node, host);
-
-        // call or syscall are not allowed inside a syscall
-        if self.in_syscall {
-            let instruction = if call_node.is_syscall() { "syscall" } else { "call" };
-            return Err(ExecutionError::CallInSyscall(instruction));
-        }
-
-        let callee_hash = current_forest
-            .get_node_by_id(call_node.callee())
-            .ok_or(ExecutionError::MastNodeNotFoundInForest { node_id: call_node.callee() })?
-            .digest();
-
-        self.save_context_and_truncate_stack(tracer);
-
-        if call_node.is_syscall() {
-            // check if the callee is in the kernel
-            if !program.kernel().contains_proc(callee_hash) {
-                return Err(ExecutionError::syscall_target_not_in_kernel(callee_hash, &err_ctx));
-            }
-
-            // set the system registers to the syscall context
-            self.ctx = ContextId::root();
-            self.fmp = SYSCALL_FMP_MIN.into();
-            self.in_syscall = true;
-        } else {
-            // set the system registers to the callee context. The value for the new context is
-            // defined to be the value of the clock cycle for the following operation.
-            self.ctx = (self.clk + 1).into();
-            self.fmp = Felt::new(FMP_MIN);
-            self.caller_hash = callee_hash;
-        }
-
-        // push the callee onto the continuation stack, and increment the clock (corresponding to
-        // the row inserted for the CALL or SYSCALL operation added to the trace).
-        continuation_stack.push_finish_call(current_node_id);
-        continuation_stack.push_start_node(call_node.callee());
-
-        // Corresponds to the CALL or SYSCALL operation added to the trace.
-        self.increment_clk(tracer);
-
-        Ok(())
-    }
-
-    /// Executes the finish phase of a Call node.
-    #[inline(always)]
-    fn finish_call_node(
-        &mut self,
-        node_id: MastNodeId,
-        current_forest: &Arc<MastForest>,
-        continuation_stack: &mut ContinuationStack,
-        host: &mut impl AsyncHost,
-        tracer: &mut impl Tracer,
-    ) -> Result<(), ExecutionError> {
-        tracer.start_clock_cycle(
-            self,
-            NodeExecutionState::End(node_id),
-            continuation_stack,
-            current_forest,
-        );
-
-        let call_node = current_forest[node_id].unwrap_call();
-        let err_ctx = err_ctx!(current_forest, call_node, host);
-        // when returning from a function call or a syscall, restore the
-        // context of the
-        // system registers and the operand stack to what it was prior
-        // to the call.
-        self.restore_context(tracer, &err_ctx)?;
-
-        // Corresponds to the row inserted for the END operation added to the trace.
-        self.increment_clk(tracer);
-        self.execute_after_exit_decorators(node_id, current_forest, host)
-    }
-
-    /// Executes a Dyn node from the start.
-    #[inline(always)]
-    async fn start_dyn_node(
-        &mut self,
-        current_node_id: MastNodeId,
-        current_forest: &mut Arc<MastForest>,
-        continuation_stack: &mut ContinuationStack,
-        host: &mut impl AsyncHost,
-        tracer: &mut impl Tracer,
-    ) -> Result<(), ExecutionError> {
-        tracer.start_clock_cycle(
-            self,
-            NodeExecutionState::Start(current_node_id),
-            continuation_stack,
-            current_forest,
-        );
-
-        // Execute decorators that should be executed before entering the node
-        self.execute_before_enter_decorators(current_node_id, current_forest, host)?;
-
-        // Corresponds to the row inserted for the DYN or DYNCALL operation
-        // added to the trace.
-        let dyn_node = current_forest[current_node_id].unwrap_dyn();
-
-        // dyn calls are not allowed inside a syscall
-        if dyn_node.is_dyncall() && self.in_syscall {
-            return Err(ExecutionError::CallInSyscall("dyncall"));
-        }
-
-        let err_ctx = err_ctx!(&current_forest, dyn_node, host);
-
-        // Retrieve callee hash from memory, using stack top as the memory
-        // address.
-        let callee_hash = {
-            let mem_addr = self.stack_get(0);
-            self.memory
-                .read_word(self.ctx, mem_addr, self.clk, &err_ctx, tracer)
-                .map_err(ExecutionError::MemoryError)?
-        };
-
-        // Drop the memory address from the stack. This needs to be done before saving the context.
-        self.decrement_stack_size(tracer);
-
-        // For dyncall, save the context and reset it.
-        if dyn_node.is_dyncall() {
-            self.save_context_and_truncate_stack(tracer);
-            // The value for the new context is defined to be the value of the clock cycle for the
-            // following operation.
-            self.ctx = (self.clk + 1).into();
-            self.fmp = Felt::new(FMP_MIN);
-            self.caller_hash = callee_hash;
-        };
-
-        // Update continuation stack
-        // -----------------------------
-        continuation_stack.push_finish_dyn(current_node_id);
-
-        // if the callee is not in the program's MAST forest, try to find a MAST forest for it in
-        // the host (corresponding to an external library loaded in the host); if none are found,
-        // return an error.
-        match current_forest.find_procedure_root(callee_hash) {
-            Some(callee_id) => {
-                continuation_stack.push_start_node(callee_id);
-            },
-            None => {
-                let (root_id, new_forest) = self
-                    .load_mast_forest(
-                        callee_hash,
-                        host,
-                        ExecutionError::dynamic_node_not_found,
-                        &err_ctx,
-                    )
-                    .await?;
-
-                // Push current forest to the continuation stack so that we can return to it
-                continuation_stack.push_enter_forest(current_forest.clone());
-
-                // Push the root node of the external MAST forest onto the continuation stack.
-                continuation_stack.push_start_node(root_id);
-
-                // Set the new MAST forest as current
-                *current_forest = new_forest;
-            },
-        }
-
-        // Increment the clock, corresponding to the row inserted for the DYN or DYNCALL operation
-        // added to the trace.
-        self.increment_clk(tracer);
-
-        Ok(())
-    }
-
-    /// Executes the finish phase of a Dyn node.
-    #[inline(always)]
-    fn finish_dyn_node(
-        &mut self,
-        node_id: MastNodeId,
-        current_forest: &Arc<MastForest>,
-        continuation_stack: &mut ContinuationStack,
-        host: &mut impl AsyncHost,
-        tracer: &mut impl Tracer,
-    ) -> Result<(), ExecutionError> {
-        tracer.start_clock_cycle(
-            self,
-            NodeExecutionState::End(node_id),
-            continuation_stack,
-            current_forest,
-        );
-
-        let dyn_node = current_forest[node_id].unwrap_dyn();
-        let err_ctx = err_ctx!(current_forest, dyn_node, host);
-        // For dyncall, restore the context.
-        if dyn_node.is_dyncall() {
-            self.restore_context(tracer, &err_ctx)?;
-        }
-
-        // Corresponds to the row inserted for the END operation added to
-        // the trace.
-        self.increment_clk(tracer);
-        self.execute_after_exit_decorators(node_id, current_forest, host)
-    }
-
-    /// Executes an External node.
-    #[inline(always)]
-    async fn execute_external_node(
-        &mut self,
-        external_node_id: MastNodeId,
-        current_forest: &mut Arc<MastForest>,
-        continuation_stack: &mut ContinuationStack,
-        host: &mut impl AsyncHost,
-        tracer: &mut impl Tracer,
-    ) -> Result<(), ExecutionError> {
-        // Execute decorators that should be executed before entering the node
-        self.execute_before_enter_decorators(external_node_id, current_forest, host)?;
-
-        let external_node = current_forest[external_node_id].unwrap_external();
-        let (resolved_node_id, new_mast_forest) =
-            self.resolve_external_node(external_node, host).await?;
-
-        tracer.record_external_node_resolution(resolved_node_id, &new_mast_forest);
-
-        // Push current forest to the continuation stack so that we can return to it
-        continuation_stack.push_enter_forest(current_forest.clone());
-
-        // Push the root node of the external MAST forest onto the continuation stack.
-        continuation_stack.push_start_node(resolved_node_id);
-
-        self.execute_after_exit_decorators(external_node_id, current_forest, host)?;
-
-        // Update the current forest to the new MAST forest.
-        *current_forest = new_mast_forest;
-
-        Ok(())
-    }
-
-    // Note: when executing individual ops, we do not increment the clock by 1 at every iteration
-    // for performance reasons (~25% performance drop). Hence, `self.clk` cannot be used directly to
-    // determine the number of operations executed in a program.
-    #[allow(clippy::too_many_arguments)]
-    #[inline(always)]
-    async fn execute_basic_block_node(
-        &mut self,
-        basic_block_node: &BasicBlockNode,
-        node_id: MastNodeId,
-        program: &MastForest,
-        host: &mut impl AsyncHost,
-        continuation_stack: &mut ContinuationStack,
-        current_forest: &Arc<MastForest>,
-        tracer: &mut impl Tracer,
-    ) -> Result<(), ExecutionError> {
-        tracer.start_clock_cycle(
-            self,
-            NodeExecutionState::Start(node_id),
-            continuation_stack,
-            current_forest,
-        );
-
-        // Execute decorators that should be executed before entering the node
-        self.execute_before_enter_decorators(node_id, program, host)?;
-
-        // Corresponds to the row inserted for the SPAN operation added to the trace.
-        self.increment_clk(tracer);
-
-        let mut batch_offset_in_block = 0;
-        let mut op_batches = basic_block_node.op_batches().iter();
-        let mut decorator_ids = basic_block_node.decorator_iter();
-
-        // execute first op batch
-        if let Some(first_op_batch) = op_batches.next() {
-            self.execute_op_batch(
-                basic_block_node,
-                node_id,
-                first_op_batch,
-                0,
-                &mut decorator_ids,
-                batch_offset_in_block,
-                program,
-                host,
-                continuation_stack,
-                current_forest,
-                tracer,
-            )
-            .await?;
-            batch_offset_in_block += first_op_batch.ops().len();
-        }
-
-        // execute the rest of the op batches
-        for (batch_index_minus_1, op_batch) in op_batches.enumerate() {
-            // RESPAN
-            {
-                tracer.start_clock_cycle(
-                    self,
-                    NodeExecutionState::Respan {
-                        node_id,
-                        batch_index: batch_index_minus_1 + 1,
-                    },
-                    continuation_stack,
-                    current_forest,
-                );
-
-                // Corresponds to the RESPAN operation added to the trace.
-                self.increment_clk(tracer);
-            }
-
-            self.execute_op_batch(
-                basic_block_node,
-                node_id,
-                op_batch,
-                batch_index_minus_1 + 1,
-                &mut decorator_ids,
-                batch_offset_in_block,
-                program,
-                host,
-                continuation_stack,
-                current_forest,
-                tracer,
-            )
-            .await?;
-            batch_offset_in_block += op_batch.ops().len();
-        }
-
-        tracer.start_clock_cycle(
-            self,
-            NodeExecutionState::End(node_id),
-            continuation_stack,
-            current_forest,
-        );
-
-        // Corresponds to the row inserted for the END operation added to the trace.
-        self.increment_clk(tracer);
-
-        // execute any decorators which have not been executed during span ops execution; this can
-        // happen for decorators appearing after all operations in a block. these decorators are
-        // executed after SPAN block is closed to make sure the VM clock cycle advances beyond the
-        // last clock cycle of the SPAN block ops.
-        for &decorator_id in decorator_ids {
-            let decorator = program
-                .get_decorator_by_id(decorator_id)
-                .ok_or(ExecutionError::DecoratorNotFoundInForest { decorator_id })?;
-            self.execute_decorator(decorator, host)?;
-        }
-
-        self.execute_after_exit_decorators(node_id, program, host)
-    }
-
-    #[inline(always)]
-    #[allow(clippy::too_many_arguments)]
-    async fn execute_op_batch(
-        &mut self,
-        basic_block: &BasicBlockNode,
-        node_id: MastNodeId,
-        batch: &OpBatch,
-        batch_index: usize,
-        decorators: &mut DecoratorIterator<'_>,
-        batch_offset_in_block: usize,
-        program: &MastForest,
-        host: &mut impl AsyncHost,
-        continuation_stack: &mut ContinuationStack,
-        current_forest: &Arc<MastForest>,
-        tracer: &mut impl Tracer,
-    ) -> Result<(), ExecutionError> {
-        let end_indices = batch.end_indices();
-        let mut group_idx = 0;
-        let mut next_group_idx = 1;
-
-        // execute operations in the batch one by one
-        for (op_idx_in_batch, op) in batch.ops().iter().enumerate() {
-            let op_idx_in_block = batch_offset_in_block + op_idx_in_batch;
-            while let Some(&decorator_id) = decorators.next_filtered(op_idx_in_block) {
-                let decorator = program
-                    .get_decorator_by_id(decorator_id)
-                    .ok_or(ExecutionError::DecoratorNotFoundInForest { decorator_id })?;
-                self.execute_decorator(decorator, host)?;
-            }
-
-            // if in trace mode, check if we need to record a trace state before executing the
-            // operation
-            tracer.start_clock_cycle(
-                self,
-                NodeExecutionState::BasicBlock { node_id, batch_index, op_idx_in_batch },
-                continuation_stack,
-                current_forest,
-            );
-
-            let err_ctx = err_ctx!(program, basic_block, host, op_idx_in_block);
-
-            // Execute the operation.
-            //
-            // Note: we handle the `Emit` operation separately, because it is an async operation,
-            // whereas all the other operations are synchronous (resulting in a significant
-            // performance improvement).
-            match op {
-                Operation::Emit => {
-                    if self.bounds_check_counter == 0 {
-                        let err_str = if self.stack_top_idx - MIN_STACK_DEPTH == 0 {
-                            "stack underflow"
-                        } else {
-                            "stack overflow"
-                        };
-                        return Err(ExecutionError::FailedToExecuteProgram(err_str));
-                    }
-
-                    self.op_emit(host, &err_ctx).await?
-                },
-                _ => {
-                    // if the operation is not an Emit, we execute it normally
-                    self.execute_op(op, op_idx_in_block, program, host, &err_ctx, tracer)?;
-                },
-            }
-
-            // if the operation carries an immediate value, the value is stored at the next group
-            // pointer; so, we advance the pointer to the following group
-            let has_imm = op.imm_value().is_some();
-            if has_imm {
-                next_group_idx += 1;
-            }
-
-            // determine if we've executed all operations in a group
-            if op_idx_in_batch + 1 == end_indices[group_idx] {
-                // then, move to the next group and reset operation index
-                group_idx = next_group_idx;
-                next_group_idx += 1;
-            }
-
-            self.increment_clk(tracer);
-        }
-
-        Ok(())
-    }
 
     /// Executes the decorators that should be executed before entering a node.
     fn execute_before_enter_decorators(
@@ -1238,151 +545,6 @@ impl FastProcessor {
         Ok(())
     }
 
-    /// Executes the given operation.
-    ///
-    /// # Panics
-    /// - if the operation is a control flow operation, as these are never executed,
-    /// - if the operation is an `Emit` operation, as this requires async execution.
-    #[inline(always)]
-    fn execute_op(
-        &mut self,
-        operation: &Operation,
-        op_idx: usize,
-        program: &MastForest,
-        host: &mut impl AsyncHost,
-        err_ctx: &impl ErrorContext,
-        tracer: &mut impl Tracer,
-    ) -> Result<(), ExecutionError> {
-        if self.bounds_check_counter == 0 {
-            let err_str = if self.stack_top_idx - MIN_STACK_DEPTH == 0 {
-                "stack underflow"
-            } else {
-                "stack overflow"
-            };
-            return Err(ExecutionError::FailedToExecuteProgram(err_str));
-        }
-
-        match operation {
-            // ----- system operations ------------------------------------------------------------
-            Operation::Noop => {
-                // do nothing
-            },
-            Operation::Assert(err_code) => {
-                self.op_assert(*err_code, host, program, err_ctx, tracer)?
-            },
-            Operation::FmpAdd => self.op_fmpadd(),
-            Operation::FmpUpdate => self.op_fmpupdate(tracer)?,
-            Operation::SDepth => self.op_sdepth(tracer),
-            Operation::Caller => self.op_caller()?,
-            Operation::Clk => self.op_clk(tracer)?,
-            Operation::Emit => {
-                panic!("emit instruction requires async, so is not supported by execute_op()")
-            },
-
-            // ----- flow control operations ------------------------------------------------------
-            // control flow operations are never executed directly
-            Operation::Join => unreachable!("control flow operation"),
-            Operation::Split => unreachable!("control flow operation"),
-            Operation::Loop => unreachable!("control flow operation"),
-            Operation::Call => unreachable!("control flow operation"),
-            Operation::SysCall => unreachable!("control flow operation"),
-            Operation::Dyn => unreachable!("control flow operation"),
-            Operation::Dyncall => unreachable!("control flow operation"),
-            Operation::Span => unreachable!("control flow operation"),
-            Operation::Repeat => unreachable!("control flow operation"),
-            Operation::Respan => unreachable!("control flow operation"),
-            Operation::End => unreachable!("control flow operation"),
-            Operation::Halt => unreachable!("control flow operation"),
-
-            // ----- field operations -------------------------------------------------------------
-            Operation::Add => self.op_add(tracer)?,
-            Operation::Neg => self.op_neg()?,
-            Operation::Mul => self.op_mul(tracer)?,
-            Operation::Inv => self.op_inv(err_ctx)?,
-            Operation::Incr => self.op_incr()?,
-            Operation::And => self.op_and(tracer, err_ctx)?,
-            Operation::Or => self.op_or(tracer, err_ctx)?,
-            Operation::Not => self.op_not(err_ctx)?,
-            Operation::Eq => self.op_eq(tracer)?,
-            Operation::Eqz => self.op_eqz()?,
-            Operation::Expacc => self.op_expacc(),
-            Operation::Ext2Mul => self.op_ext2mul(),
-
-            // ----- u32 operations ---------------------------------------------------------------
-            Operation::U32split => self.op_u32split(tracer),
-            Operation::U32add => self.op_u32add(err_ctx)?,
-            Operation::U32add3 => self.op_u32add3(err_ctx, tracer)?,
-            Operation::U32sub => self.op_u32sub(op_idx, err_ctx, tracer)?,
-            Operation::U32mul => self.op_u32mul(err_ctx)?,
-            Operation::U32madd => self.op_u32madd(err_ctx, tracer)?,
-            Operation::U32div => self.op_u32div(err_ctx, tracer)?,
-            Operation::U32and => self.op_u32and(err_ctx, tracer)?,
-            Operation::U32xor => self.op_u32xor(err_ctx, tracer)?,
-            Operation::U32assert2(err_code) => self.op_u32assert2(*err_code, err_ctx, tracer)?,
-
-            // ----- stack manipulation -----------------------------------------------------------
-            Operation::Pad => self.op_pad(tracer),
-            Operation::Drop => self.decrement_stack_size(tracer),
-            Operation::Dup0 => self.dup_nth(0, tracer),
-            Operation::Dup1 => self.dup_nth(1, tracer),
-            Operation::Dup2 => self.dup_nth(2, tracer),
-            Operation::Dup3 => self.dup_nth(3, tracer),
-            Operation::Dup4 => self.dup_nth(4, tracer),
-            Operation::Dup5 => self.dup_nth(5, tracer),
-            Operation::Dup6 => self.dup_nth(6, tracer),
-            Operation::Dup7 => self.dup_nth(7, tracer),
-            Operation::Dup9 => self.dup_nth(9, tracer),
-            Operation::Dup11 => self.dup_nth(11, tracer),
-            Operation::Dup13 => self.dup_nth(13, tracer),
-            Operation::Dup15 => self.dup_nth(15, tracer),
-            Operation::Swap => self.op_swap(),
-            Operation::SwapW => self.swapw_nth(1),
-            Operation::SwapW2 => self.swapw_nth(2),
-            Operation::SwapW3 => self.swapw_nth(3),
-            Operation::SwapDW => self.op_swap_double_word(),
-            Operation::MovUp2 => self.rotate_left(3),
-            Operation::MovUp3 => self.rotate_left(4),
-            Operation::MovUp4 => self.rotate_left(5),
-            Operation::MovUp5 => self.rotate_left(6),
-            Operation::MovUp6 => self.rotate_left(7),
-            Operation::MovUp7 => self.rotate_left(8),
-            Operation::MovUp8 => self.rotate_left(9),
-            Operation::MovDn2 => self.rotate_right(3),
-            Operation::MovDn3 => self.rotate_right(4),
-            Operation::MovDn4 => self.rotate_right(5),
-            Operation::MovDn5 => self.rotate_right(6),
-            Operation::MovDn6 => self.rotate_right(7),
-            Operation::MovDn7 => self.rotate_right(8),
-            Operation::MovDn8 => self.rotate_right(9),
-            Operation::CSwap => self.op_cswap(err_ctx, tracer)?,
-            Operation::CSwapW => self.op_cswapw(err_ctx, tracer)?,
-
-            // ----- input / output ---------------------------------------------------------------
-            Operation::Push(element) => self.op_push(*element, tracer),
-            Operation::AdvPop => self.op_advpop(err_ctx, tracer)?,
-            Operation::AdvPopW => self.op_advpopw(err_ctx, tracer)?,
-            Operation::MLoadW => self.op_mloadw(err_ctx, tracer)?,
-            Operation::MStoreW => self.op_mstorew(err_ctx, tracer)?,
-            Operation::MLoad => self.op_mload(err_ctx, tracer)?,
-            Operation::MStore => self.op_mstore(err_ctx, tracer)?,
-            Operation::MStream => self.op_mstream(err_ctx, tracer)?,
-            Operation::Pipe => self.op_pipe(err_ctx, tracer)?,
-
-            // ----- cryptographic operations -----------------------------------------------------
-            Operation::HPerm => self.op_hperm(tracer),
-            Operation::MpVerify(err_code) => {
-                self.op_mpverify(*err_code, program, tracer, err_ctx)?
-            },
-            Operation::MrUpdate => self.op_mrupdate(tracer, err_ctx)?,
-            Operation::FriE2F4 => self.op_fri_ext2fold4(tracer)?,
-            Operation::HornerBase => self.op_horner_eval_base(tracer, err_ctx)?,
-            Operation::HornerExt => self.op_horner_eval_ext(tracer, err_ctx)?,
-            Operation::EvalCircuit => self.op_eval_circuit(err_ctx)?,
-        }
-
-        Ok(())
-    }
-
     // HELPERS
     // ----------------------------------------------------------------------------------------------
 
@@ -1427,31 +589,6 @@ impl FastProcessor {
         Ok((root_id, mast_forest))
     }
 
-    /// Analogous to [`Process::resolve_external_node`](crate::Process::resolve_external_node), but
-    /// for asynchronous execution.
-    async fn resolve_external_node(
-        &mut self,
-        external_node: &ExternalNode,
-        host: &mut impl AsyncHost,
-    ) -> Result<(MastNodeId, Arc<MastForest>), ExecutionError> {
-        let (root_id, mast_forest) = self
-            .load_mast_forest(
-                external_node.digest(),
-                host,
-                ExecutionError::no_mast_forest_with_procedure,
-                &(),
-            )
-            .await?;
-
-        // if the node that we got by looking up an external reference is also an External
-        // node, we are about to enter into an infinite loop - so, return an error
-        if mast_forest[root_id].is_external() {
-            return Err(ExecutionError::CircularExternalNode(external_node.digest()));
-        }
-
-        Ok((root_id, mast_forest))
-    }
-
     /// Increments the stack top pointer by 1.
     ///
     /// The bottom of the stack is never affected by this operation.
@@ -1474,12 +611,6 @@ impl FastProcessor {
         self.stack_top_idx -= 1;
         self.stack_bot_idx = min(self.stack_bot_idx, self.stack_top_idx - MIN_STACK_DEPTH);
         self.update_bounds_check_counter();
-    }
-
-    /// Returns the size of the stack.
-    #[inline(always)]
-    fn stack_size(&self) -> usize {
-        self.stack_top_idx - self.stack_bot_idx
     }
 
     /// Updates the bounds check counter.
@@ -1506,83 +637,6 @@ impl FastProcessor {
             self.bounds_check_counter =
                 min(self.stack_top_idx - MIN_STACK_DEPTH, STACK_BUFFER_SIZE - self.stack_top_idx);
         }
-    }
-
-    /// Saves the current execution context and truncates the stack to 16 elements in preparation to
-    /// start a new execution context.
-    fn save_context_and_truncate_stack(&mut self, tracer: &mut impl Tracer) {
-        let overflow_stack = if self.stack_size() > MIN_STACK_DEPTH {
-            // save the overflow stack, and zero out the buffer.
-            //
-            // Note: we need to zero the overflow buffer, since the new context expects ZERO's to be
-            // pulled in if they decrement the stack size (e.g. by executing a `drop`).
-            let overflow_stack =
-                self.stack[self.stack_bot_idx..self.stack_top_idx - MIN_STACK_DEPTH].to_vec();
-            self.stack[self.stack_bot_idx..self.stack_top_idx - MIN_STACK_DEPTH].fill(ZERO);
-
-            overflow_stack
-        } else {
-            Vec::new()
-        };
-
-        self.stack_bot_idx = self.stack_top_idx - MIN_STACK_DEPTH;
-
-        self.call_stack.push(ExecutionContextInfo {
-            overflow_stack,
-            ctx: self.ctx,
-            fn_hash: self.caller_hash,
-            fmp: self.fmp,
-        });
-
-        tracer.start_context();
-    }
-
-    /// Restores the execution context to the state it was in before the last `call`, `syscall` or
-    /// `dyncall`.
-    ///
-    /// This includes restoring the overflow stack and the system parameters.
-    ///
-    /// # Errors
-    /// - Returns an error if the overflow stack is larger than the space available in the stack
-    ///   buffer.
-    fn restore_context(
-        &mut self,
-        tracer: &mut impl Tracer,
-        err_ctx: &impl ErrorContext,
-    ) -> Result<(), ExecutionError> {
-        // when a call/dyncall/syscall node ends, stack depth must be exactly 16.
-        if self.stack_size() > MIN_STACK_DEPTH {
-            return Err(ExecutionError::invalid_stack_depth_on_return(self.stack_size(), err_ctx));
-        }
-
-        let ctx_info = self
-            .call_stack
-            .pop()
-            .expect("execution context stack should never be empty when restoring context");
-
-        // restore the overflow stack
-        {
-            let overflow_len = ctx_info.overflow_stack.len();
-            if overflow_len > self.stack_bot_idx {
-                return Err(ExecutionError::FailedToExecuteProgram(
-                    "stack underflow when restoring context",
-                ));
-            }
-
-            self.stack[range(self.stack_bot_idx - overflow_len, overflow_len)]
-                .copy_from_slice(&ctx_info.overflow_stack);
-            self.stack_bot_idx -= overflow_len;
-        }
-
-        // restore system parameters
-        self.ctx = ctx_info.ctx;
-        self.fmp = ctx_info.fmp;
-        self.in_syscall = false;
-        self.caller_hash = ctx_info.fn_hash;
-
-        tracer.restore_context();
-
-        Ok(())
     }
 
     // TESTING
