+++ stdlib/docs/sys/vm/mod.md
@@ -1,0 +1,5 @@
+
+## std::sys::vm::mod
+| Procedure | Description |
+| ----------- | ------------- |
+| verify_proof | Verifies a STARK proof attesting to the correct execution of a program in the Miden VM.<br /><br />- The public inputs are composed of the input and output stacks, of fixed size equal to 16, as<br />well as the program and the kernel procedures digests.<br />- There are two trace segments, main and auxiliary. It is assumed that the main trace segment<br />is 73 columns wide while the auxiliary trace segment is 8 columns wide. Note that we pad the main<br />trace to the next multiple of 8.<br />- The OOD evaluation frame is composed of two concatenated rows, current and next, each composed<br />of 73 elements representing the main trace portion and 8 elements for the auxiliary trace one.<br />Note that, due to the padding of the main trace columns, the number of OOD evaluations per row<br />is 80 for the main trace.<br /><br />Inputs:  [log(trace_length), num_queries, grinding]<br />Outputs: []<br /><br />Cycles:<br />1- Remainder polynomial size 64:<br />2515 + num_queries * (512 + num_fri_layers * 83) + 108 * num_fri_layers + 10 * log(trace_length)<br />2- Remainder polynomial size 128:<br />2540 + num_queries * (541 + num_fri_layers * 83) + 108 * num_fri_layers + 10 * log(trace_length)<br /><br />where num_fri_layers is computed as:<br /><br />1- If log(trace_length) is even, then num_fri_layers = (log(trace_length) - 6) / 2, where 6 = log2(64),<br />2- If log(trace_length) is odd, then num_fri_layers = (log(trace_length) - 7) / 2, where 7 = log2(128).<br /> |
